requirements.txt
selenium==4.9.0
webdriver-manager
pytest
requests
allure-pytest==2.9.45
jproperties
jsonschema